test(skills): split reviewer-mode workshop tool tests into their own file
