+++ src/agents/tools/skill-workshop-tool.review.test.ts
@@ -1,0 +1,295 @@
+// skill_workshop review-mode tests cover the proposal-only reviewer surface:
+// mutation budgets, read receipts, and patch/update drafting for live skills.
+import fs from "node:fs/promises";
+import path from "node:path";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import type { SkillWorkshopProposalMutationBudget } from "../../skills/workshop/types.js";
+import {
+  createOpenClawTestState,
+  type OpenClawTestState,
+} from "../../test-utils/openclaw-test-state.js";
+import { createTrackedTempDirs } from "../../test-utils/tracked-temp-dirs.js";
+import { createSkillWorkshopTool } from "./skill-workshop-tool.js";
+
+const tempDirs = createTrackedTempDirs();
+let testState: OpenClawTestState;
+
+beforeEach(async () => {
+  testState = await createOpenClawTestState({
+    layout: "state-only",
+    prefix: "openclaw-skill-workshop-review-state-",
+  });
+});
+
+afterEach(async () => {
+  await testState.cleanup();
+  await tempDirs.cleanup();
+});
+
+async function seedLiveSkill(
+  workspaceDir: string,
+  name: string,
+  description: string,
+  content: string,
+): Promise<void> {
+  const fullTool = createSkillWorkshopTool({
+    workspaceDir,
+    config: { skills: { workshop: { approvalPolicy: "auto" } } },
+  });
+  const created = await fullTool.execute("seed-create", {
+    action: "create",
+    name,
+    description,
+    proposal_content: content,
+  });
+  await fullTool.execute("seed-apply", {
+    action: "apply",
+    proposal_id: (created.details as { id: string }).id,
+    reason: "seed live skill",
+  });
+}
+
+describe("skill_workshop review mode", () => {
+  it("restricts internal review runs to one pending proposal mutation", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-review-");
+    const proposalMutationBudget: SkillWorkshopProposalMutationBudget = { remaining: 1 };
+    const tool = createSkillWorkshopTool({
+      workspaceDir,
+      config: { skills: { workshop: { approvalPolicy: "auto" } } },
+      proposalOnly: true,
+      proposalMutationBudget,
+    });
+
+    expect(
+      (tool.parameters as { properties: { action: { enum: string[] } } }).properties.action.enum,
+    ).toEqual(["create", "revise", "list", "inspect"]);
+    await expect(
+      tool.execute("call-apply", { action: "apply", proposal_id: "proposal-1" }),
+    ).rejects.toThrow("only inspect or draft proposals");
+    await expect(
+      tool.execute("call-evaluate", { action: "evaluate", proposal_id: "proposal-1" }),
+    ).rejects.toThrow("only inspect or draft proposals");
+    await expect(
+      tool.execute("call-update", {
+        action: "update",
+        skill_name: "existing-skill",
+        proposal_content: "# Replacement\n",
+      }),
+    ).rejects.toThrow("only inspect or draft proposals");
+
+    await tool.execute("call-create", {
+      action: "create",
+      name: "Review Learning",
+      description: "Reuse a recovered workflow",
+      proposal_content: "# Review Learning\n\nFollow the recovered workflow.\n",
+    });
+    expect(proposalMutationBudget.completed).toBe(1);
+    const retryTool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      proposalMutationBudget,
+    });
+    await expect(
+      retryTool.execute("call-create-2", {
+        action: "create",
+        name: "Second Learning",
+        description: "Should stay blocked",
+        proposal_content: "# Second Learning\n",
+      }),
+    ).rejects.toThrow("reached its proposal mutation limit");
+  });
+
+  it("lets internal review runs draft update proposals for existing skills", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-review-update-");
+    await seedLiveSkill(
+      workspaceDir,
+      "weather-planner",
+      "Plan around the weather forecast",
+      "# Weather Planner\n\nCheck weather before outdoor recommendations.\n",
+    );
+
+    const proposalMutationBudget: SkillWorkshopProposalMutationBudget = { remaining: 1 };
+    const reviewTool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      updateProposals: true,
+      proposalMutationBudget,
+    });
+    const update = await reviewTool.execute("review-update", {
+      action: "update",
+      skill_name: "weather-planner",
+      proposal_content:
+        "# Weather Planner\n\nCheck weather before outdoor recommendations.\nCheck alerts and timing.\n",
+    });
+
+    expect(update.details).toMatchObject({
+      status: "pending",
+      kind: "update",
+      skillKey: "weather-planner",
+    });
+    expect(proposalMutationBudget.remaining).toBe(0);
+  });
+
+  it("composes patch proposals by replacing the quoted span of the live body", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-review-extend-");
+    await seedLiveSkill(
+      workspaceDir,
+      "weather-planner",
+      "Plan around the weather forecast",
+      "# Weather Planner\n\nCheck weather before outdoor recommendations.\n",
+    );
+
+    const proposalMutationBudget: SkillWorkshopProposalMutationBudget = { remaining: 1 };
+    const reviewTool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      updateProposals: true,
+      proposalMutationBudget,
+    });
+    expect(
+      (reviewTool.parameters as { properties: { action: { enum: string[] } } }).properties.action
+        .enum,
+    ).toEqual(["create", "patch", "update", "read", "revise", "list", "inspect"]);
+
+    await expect(
+      reviewTool.execute("patch-without-read", {
+        action: "patch",
+        skill_name: "weather-planner",
+        old_string: "Check weather before outdoor recommendations.",
+        new_string: "Replacement.",
+      }),
+    ).rejects.toThrow("read the live skill first");
+
+    await reviewTool.execute("review-read", { action: "read", skill_name: "weather-planner" });
+    await expect(
+      reviewTool.execute("patch-no-match", {
+        action: "patch",
+        skill_name: "weather-planner",
+        old_string: "Text that is not in the skill.",
+        new_string: "Replacement.",
+      }),
+    ).rejects.toThrow("not found in the live skill body");
+
+    const extended = await reviewTool.execute("review-patch", {
+      action: "patch",
+      skill_name: "weather-planner",
+      old_string: "Check weather before outdoor recommendations.",
+      new_string:
+        "Check weather before outdoor recommendations.\nCheck alerts and timing before recommending.",
+    });
+
+    expect(extended.details).toMatchObject({
+      status: "pending",
+      kind: "update",
+      skillKey: "weather-planner",
+    });
+    expect(proposalMutationBudget.patchProposalIds?.size).toBe(1);
+    const inspected = await createSkillWorkshopTool({ workspaceDir }).execute("inspect-extend", {
+      action: "inspect",
+      proposal_id: (extended.details as { id: string }).id,
+    });
+    const content = (inspected as { details: { content?: string } }).details.content ?? "";
+    const inspectText = (inspected.content[0] as { text: string }).text;
+    const proposalBody = content || inspectText;
+    expect(proposalBody).toContain("Check weather before outdoor recommendations.");
+    expect(proposalBody).toContain("Check alerts and timing before recommending.");
+  });
+
+  it("refuses a patch when the skill changed after the read", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-stale-patch-");
+    await seedLiveSkill(
+      workspaceDir,
+      "weather-planner",
+      "Plan around the weather forecast",
+      "# Weather Planner\n\nCheck weather before outdoor recommendations.\n",
+    );
+
+    const proposalMutationBudget: SkillWorkshopProposalMutationBudget = { remaining: 1 };
+    const reviewTool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      updateProposals: true,
+      proposalMutationBudget,
+    });
+    await reviewTool.execute("review-read", { action: "read", skill_name: "weather-planner" });
+    const liveSkillFile = path.join(workspaceDir, "skills", "weather-planner", "SKILL.md");
+    await fs.writeFile(
+      liveSkillFile,
+      (await fs.readFile(liveSkillFile, "utf8")).replace(
+        "Check weather before outdoor recommendations.",
+        "Operator-edited steps after the read.",
+      ),
+    );
+    await expect(
+      reviewTool.execute("stale-patch", {
+        action: "patch",
+        skill_name: "weather-planner",
+        old_string: "Check weather before outdoor recommendations.",
+        new_string: "Replacement.",
+      }),
+    ).rejects.toThrow("changed since it was read");
+    expect(proposalMutationBudget.remaining).toBe(1);
+  });
+
+  it("caps reviewer live-skill reads at the read budget", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-review-read-cap-");
+    await seedLiveSkill(
+      workspaceDir,
+      "big-skill",
+      "A very large operator skill",
+      `# Big Skill\n\n${"A detailed operational line.\n".repeat(1200)}`,
+    );
+
+    const reviewTool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      updateProposals: true,
+      proposalMutationBudget: { remaining: 1 },
+    });
+    const read = await reviewTool.execute("review-read", {
+      action: "read",
+      skill_name: "big-skill",
+    });
+    const text = (read.content[0] as { text: string }).text;
+    expect(read.details).toMatchObject({ skillKey: "big-skill", truncated: true });
+    expect(text.length).toBeLessThanOrEqual(20_000 + 100);
+    expect(text).toContain("[truncated: skill exceeds the reviewer read budget]");
+
+    await expect(
+      reviewTool.execute("oversized-patch", {
+        action: "patch",
+        skill_name: "big-skill",
+        old_string: "A detailed operational line.",
+        new_string: "A rewritten operational line.",
+      }),
+    ).rejects.toThrow("cannot be patched autonomously");
+  });
+
+  it("does not refund the review mutation budget after a failed mutation", async () => {
+    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-review-failure-");
+    const proposalMutationBudget: SkillWorkshopProposalMutationBudget = { remaining: 1 };
+    const tool = createSkillWorkshopTool({
+      workspaceDir,
+      proposalOnly: true,
+      proposalMutationBudget,
+    });
+
+    await expect(
+      tool.execute("call-revise-missing", {
+        action: "revise",
+        proposal_id: "missing-proposal",
+        proposal_content: "# Missing Skill\n",
+      }),
+    ).rejects.toThrow();
+    await expect(
+      tool.execute("call-create-after-failure", {
+        action: "create",
+        name: "Second Mutation",
+        description: "Must remain blocked after a failed mutation",
+        proposal_content: "# Second Mutation\n",
+      }),
+    ).rejects.toThrow("reached its proposal mutation limit");
+    expect(proposalMutationBudget.completed).toBeUndefined();
+    expect(proposalMutationBudget.failedMutations).toBe(1);
+  });
+});
